requirements.txt
fastapi
uvicorn
strawberry-graphql[fastapi]
google-cloud-firestore
firebase-admin
python-dotenv
httpx
pydantic
openai
pypdf
python-multipart
